requirements.txt
moment==0.8.2
allure_python_commons==2.8.5
pytest==5.1.3
selenium==3.141.0
